Expose sqrt on Integer via internal module
