+++ arithmoi.cabal
@@ -62,6 +62,7 @@ library
                           Math.NumberTheory.GCD.LowLevel
                           Math.NumberTheory.Powers
                           Math.NumberTheory.Powers.Squares
+                          Math.NumberTheory.Powers.Squares.Internal
                           Math.NumberTheory.Powers.Cubes
                           Math.NumberTheory.Powers.Fourth
                           Math.NumberTheory.Powers.General
